[Feat] 행사 등록 API #SCRUM-88 ##DONE

+++ ficket-event/src/main/java/com/example/ficketevent/domain/event/dto/response/SeatResponse.java
@@ -7,9 +7,9 @@ public class SeatResponse {
 
     private Long seatId; // 좌석 ID
 
-    private Integer x; // 좌표 x
+    private Double x; // 좌표 x
 
-    private Integer y; // 좌표 y
+    private Double y; // 좌표 y
 
     private String seatCol; // 열번호
